Fixes for lint
